# backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Taskosaur API",
    "description": "API for Taskosaur - Task Management Application",
    "version": "1.0.0",
    "contact": {
      "email": "info@taskosaur.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Development server"
    }
  ],
...
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "User login",
        "description": "Authenticate a user and return a token",
        "operationId": "login",
        "requestBody": {
          "description": "Login credentials",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "User registration",
        "description": "Register a new user",
        "operationId": "register",
        "requestBody": {
          "description": "User registration details",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "409": {
            "description": "Email already in use"
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["auth"],
        "summary": "Refresh token",
        "description": "Refresh authentication token",
        "operationId": "refreshToken",
        "requestBody": {
          "description": "Refresh token",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid refresh token"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["auth"],
        "summary": "Logout",
        "description": "Invalidate the user's refresh token",
        "operationId": "logout",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Logged out successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/auth/reset-password": {
      "post": {
        "tags": ["auth"],
        "summary": "Request password reset",
        "description": "Request a password reset email",
        "operationId": "requestPasswordReset",
        "requestBody": {
          "description": "Email address",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset email sent"
          },
          "404": {
            "description": "Email not found"
          }
        }
      }
    },
    "/auth/reset-password/{token}": {
      "post": {
        "tags": ["auth"],
        "summary": "Reset password",
        "description": "Reset password using a token",
        "operationId": "resetPassword",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "description": "Password reset token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "New password",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully"
          },
          "400": {
            "description": "Invalid or expired token"
          }
        }
      }
    },
...
    "/workspaces/{workspaceId}/members/{userId}": {
      "get": {
        "tags": ["roles"],
        "summary": "Get workspace member (Detailed)",
        "description": "Get details of a specific workspace member",
        "operationId": "getWorkspaceMemberDetailed",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "description": "ID of the workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace member details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceMember"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Workspace member not found"
          }
        }
      },
      "put": {
        "tags": ["roles"],
        "summary": "Update workspace member (Detailed)",
        "description": "Update a workspace member's role",
        "operationId": "updateWorkspaceMemberDetailed",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "description": "ID of the workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Member details to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceMemberRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workspace member updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceMember"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Workspace member not found"
          }
        }
      },
      "delete": {
        "tags": ["roles"],
        "summary": "Remove workspace member (Detailed)",
        "description": "Remove a member from a workspace",
        "operationId": "removeWorkspaceMemberDetailed",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "description": "ID of the workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Workspace member removed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Workspace member not found"
          }
        }
      }
    },
...
    "/projects/{projectId}/members/{userId}": {
      "get": {
        "tags": ["roles"],
        "summary": "Get project member (Detailed)",
        "description": "Get details of a specific project member",
        "operationId": "getProjectMemberDetailed",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of the project",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project member details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMember"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Project member not found"
          }
        }
      },
      "put": {
        "tags": ["roles"],
        "summary": "Update project member (Detailed)",
        "description": "Update a project member's role",
        "operationId": "updateProjectMemberDetailed",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of the project",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Member details to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectMemberRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project member updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMember"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Project member not found"
          }
        }
      },
      "delete": {
        "tags": ["roles"],
        "summary": "Remove project member (Detailed)",
        "description": "Remove a member from a project",
        "operationId": "removeProjectMemberDetailed",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of the project",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Project member removed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Project member not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "password123"
          },
          "rememberMe": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "RefreshTokenRequest": {
        "type": "object",
        "required": ["refreshToken"],
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["name", "email", "password"],
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "password123"
          }
        }
      },
      "PasswordResetRequest": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          }
        }
      },
      "NewPasswordRequest": {
        "type": "object",
        "required": ["password", "confirmPassword"],
        "properties": {
          "password": {
            "type": "string",
            "format": "password",
            "example": "newPassword123"
          },
          "confirmPassword": {
            "type": "string",
            "format": "password",
            "example": "newPassword123"
          }
        }
      },
...
      "AuthResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "user-123"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "avatar": {
            "type": "string",
            "format": "url",
            "nullable": true,
            "example": "https://example.com/avatar.png"
          },
          "role": {
            "type": "string",
            "example": "User",
            "enum": ["Admin", "User"]
          },
          "status": {
            "type": "string",
            "example": "Active",
            "enum": ["Active", "Inactive", "Pending"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-04-10T10:30:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-04-15T14:30:00Z"
          }
        }
      },
...
      "Permission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "permission-123"
          },
          "name": {
            "type": "string",
            "example": "workspace:create"
          },
          "description": {
            "type": "string",
            "example": "Permission to create workspaces"
          },
          "resource": {
            "type": "string",
            "example": "workspace"
          },
          "action": {
            "type": "string",
            "example": "create"
          }
        }
      },
      "WorkspaceMember": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "workspace-member-123"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "workspace": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "workspace-123"
              },
              "name": {
                "type": "string",
                "example": "My Workspace"
              },
              "slug": {
                "type": "string",
                "example": "my-workspace"
              }
            }
          },
          "role": {
            "type": "string",
            "example": "Owner",
            "enum": ["Owner", "Admin", "Member", "Guest"]
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-04-15T14:30:00Z"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          }
        }
      },
...
      "UpdateWorkspaceMemberRequest": {
        "type": "object",
        "required": ["role"],
        "properties": {
          "role": {
            "type": "string",
            "example": "Admin",
            "enum": ["Admin", "Member", "Guest"]
          }
        }
      },
      "ProjectMember": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "project-member-123"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "project": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "project-123"
              },
              "name": {
                "type": "string",
                "example": "My Project"
              },
              "slug": {
                "type": "string",
                "example": "my-project"
              },
              "workspace": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "example": "workspace-123"
                  },
                  "name": {
                    "type": "string",
                    "example": "My Workspace"
                  },
                  "slug": {
                    "type": "string",
                    "example": "my-workspace"
                  }
                }
              }
            }
          },
          "role": {
            "type": "string",
            "example": "Manager",
            "enum": ["Manager", "Member", "Viewer"]
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-04-15T14:30:00Z"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          }
        }
      },
...
      "UpdateProjectMemberRequest": {
        "type": "object",
        "required": ["role"],
        "properties": {
          "role": {
            "type": "string",
            "example": "Manager",
            "enum": ["Manager", "Member", "Viewer"]
          }
        }
      },
...
    }
  }
}